Add files via upload
The project was done under an academic and professional internship at the Rwanda housing authority. It took place at the head office of Rwanda Housing located in the Kacyiru sector of Gasabo district. The work was done under the supervision of Physical planner Eng.KURADUSENGE Jean Marie Vianney and Murego Theogene. 

The project was under the development of a Physical plan of one settlement in Eastern Province, Nyagatare district.